backend/requirements.txt
Django
djangorestframework
drf_social_oauth2
dj-database-url
django-heroku
django-friendship
django-filter
django-storages
django-autoslug
django-redis
django-cors-headers
whitenoise
gunicorn
coreapi
pyyaml
requests
python-dotenv

daphne
channels
channels-redis

boto
boto3
Pillow
s3transfer

